requirements.txt
numpy
PyQt5
Pillow
pydicom
opencv-python
matplotlib